Console instead of UI
Decided it would be better (and more fun) to make a console instead of a basic ui to click things. this way you can spawn items and enemies while you are dead :^]
This will also allow me to add more commands and an easy way to get help for each command and so on. Expanding with this should be much easier